games/dino: replace ulist with static arrays, drop sprintf/pow
Obstacles, clouds and the type history now live in fixed arrays inside
Horizon (worst case ~5 on-screen obstacles); ulist.c and every
malloc/free/exit call are gone, which also fixes the Obstacle/Cloud
leaks (list nodes were freed, their data never).

intToStr writes digits directly instead of dragging in sprintf;
maxScore is computed by a loop instead of pow(); minSpeed becomes
double so the pterodactyl's 8.5 threshold no longer truncates to 8
(matches Chrome).

Verified: old and new horizon produce bit-identical state traces over
20000 simulated frames with a shared RNG seed.

+++ programs/games/dino/obstacle.h
@@ -33,7 +33,7 @@ typedef struct {
     int yPosArr[3]; // used if yPos is -1
     int multipleSpeed;
     int minGap;
-    int minSpeed;
+    double minSpeed; // 8.5 for the pterodactyl, must not truncate
     int collisionBoxesCount;
     CollisionBox collisionBoxes[5];
     int numFrames;
